+++ public/data/feed_data.json
@@ -1,0 +1,38 @@
+[
+  {
+    "id": 0,
+    "profile_img": "/images/sujeongKim/profile_img.jpg",
+    "user_id": "krystal170",
+    "post_img": "/images/sujeongKim/forest.jpg",
+    "liked_user_id": "maratang77",
+    "count_liked_people": 10,
+    "reply": []
+  },
+  {
+    "id": 1,
+    "profile_img": "/images/sujeongKim/profile_img.jpg",
+    "user_id": "icecream486",
+    "post_img": "/images/sujeongKim/jellyfish_blue.jpg",
+    "liked_user_id": "dduckbboki2",
+    "count_liked_people": 7,
+    "reply": []
+  },
+  {
+    "id": 3,
+    "profile_img": "/images/sujeongKim/profile_img.jpg",
+    "user_id": "chococho543",
+    "post_img": "/images/sujeongKim/star.jpg",
+    "liked_user_id": "dimsum1234",
+    "count_liked_people": 13,
+    "reply": []
+  },
+  {
+    "id": 4,
+    "profile_img": "/images/sujeongKim/profile_img.jpg",
+    "user_id": "pizzzzza987",
+    "post_img": "/images/sujeongKim/yellow.jpg",
+    "liked_user_id": "cooooooke777",
+    "count_liked_people": 4,
+    "reply": []
+  }
+]
