+++ src/pages/sujeongKim/Main/feed.scss
@@ -1,0 +1,162 @@
+.feeds {
+  width: 820px;
+  margin: 0 35px 20px 0;
+
+  & > article {
+    margin-bottom: 50px;
+  }
+
+  & > article > header {
+    display: flex;
+    box-sizing: border-box;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    height: 80px;
+    padding: 0 20px;
+    border: 1px solid #dbdbdb;
+
+    & > div {
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+
+      & > img {
+        width: 40px;
+        border-radius: 100%;
+      }
+
+      & > p {
+        margin-left: 20px;
+        font-weight: 700;
+      }
+    }
+
+    .hiddenBtn {
+      cursor: pointer;
+    }
+  }
+
+  .postImgArea {
+    width: 820px;
+    height: 820px;
+
+    & > img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  .postBottom {
+    padding: 0 20px 5px 20px;
+    border: 1px solid #dbdbdb;
+
+    .postReaction {
+      display: flex;
+      box-sizing: border-box;
+      flex-direction: row;
+      align-items: center;
+      justify-content: space-between;
+      width: 100%;
+      height: 60px;
+      padding: 15px 0;
+
+      & > span {
+        margin-right: 10px;
+        cursor: pointer;
+      }
+
+      .likeBtn {
+        margin-right: 10px;
+        cursor: pointer;
+      }
+
+      .commentBtn {
+        margin-right: 10px;
+        cursor: pointer;
+      }
+
+      .shareBtn {
+        margin-right: 10px;
+        cursor: pointer;
+      }
+
+      .bookmarkBtn {
+        font-size: 30px;
+      }
+    }
+
+    .postLike {
+      display: flex;
+      align-items: center;
+      margin-bottom: 15px;
+
+      & > img {
+        width: 25px;
+        margin-right: 5px;
+        border: 1px solid #dbdbdb;
+        border-radius: 100%;
+      }
+
+      & > p > span {
+        font-weight: 600;
+      }
+    }
+
+    .commentList {
+      display: flex;
+      flex-direction: row;
+
+      .userIdInComment {
+        margin-right: 5px;
+        float: left;
+        font-weight: 600;
+      }
+
+      .tooLongComment {
+        display: -webkit-box;
+        width: 230px;
+        overflow: hidden;
+        font-weight: 400;
+        text-overflow: ellipsis;
+        white-space: unset;
+        -webkit-box-orient: vertical;
+        -webkit-line-clamp: 1;
+      }
+
+      .seeMore {
+        margin-left: 5px;
+        color: #999999;
+        cursor: pointer;
+      }
+      .postingTime {
+        margin-left: 10px;
+        color: #888888;
+      }
+    }
+  }
+
+  .commentInputArea {
+    display: flex;
+    justify-content: space-between;
+    height: 60px;
+    padding: 0 20px;
+    border: 1px solid #dbdbdb;
+    border-top: none;
+
+    .commentInput {
+      width: 90%;
+      border: none;
+
+      &::placeholder {
+        font-size: 15px;
+        font-weight: 400;
+      }
+    }
+
+    .commentInputBtn {
+      color: #0095f6;
+      cursor: pointer;
+    }
+  }
+}
